pyproject.toml
[tool.poetry]
name = "thesis"
version = "0.1.0"
description = ""
authors = ["Kristiyan Sakalyan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.4.1"
easydict = "^1.13"
nni = "^3.0"
networkx = {extras = ["default"], version = "^3.3"}
lightning = "^2.4.0"
torchmetrics = "^1.4.3"
numpy = "^2.1.2"
scipy = "^1.14.1"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
hydra-core = "^1.3.2"
hydra-submitit-launcher = "^1.2.0"
wandb = "^0.18.3"
tqdm = "^4.66.5"
rich = "^13.9.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
networkx = "^3.3"
matplotlib = "^3.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
